fund_accounting/requirements.txt
asgiref==3.2.7
Babel==2.5.1
certifi==2020.4.5.1
chardet==3.0.4
datetime-quarter==1.0.3
defusedxml==0.6.0
diff-match-patch==20181111
dj-database-url==0.5.0
Django==3.0.7
django-extensions==2.2.9
django-hordak==1.10.0
django-import-export==2.2.0
django-js-asset==1.2.2
django-model-utils==4.0.0
django-money==1.1
django-mptt==0.11.0
django-smalluuid==1.2.0
enum34==1.1.10
et-xmlfile==1.0.1
fancycompleter==0.9.1
idna==2.9
jdcal==1.4.1
MarkupPy==1.14
numpy==1.18.5
odfpy==1.4.1
openpyxl==3.0.3
pandas==1.0.4
pdbpp==0.10.2
psycopg2==2.8.5
psycopg2-binary==2.8.5
py-moneyed==0.8.0
Pygments==2.6.1
pyrepl==0.9.0
python-dateutil==2.8.1
pytz==2020.1
PyYAML==5.3.1
requests==2.23.0
six==1.15.0
smalluuid==1.1.0
sqlparse==0.3.1
tablib==2.0.0
urllib3==1.25.9
wmctrl==0.3
xlrd==1.2.0
xlwt==1.3.0
